Create new path for edit form with character id
